i am bad at git pushes
combined web pull, individual and group  restarts into RestartUtils so that I can make a nicer frontend later

everything is a mess though sorry

+++ IndividualRestart.ps1
@@ -1,15 +1,27 @@
 ﻿# Written by daniel angeloni
 
 while(1) {
+
     $ComputerName = Read-Host -Prompt 'Computer Name or IP Address'
-    # Attempt restart
+
+    if ($ComputerName -like "exit")
+    {
+        break
+    }
+
     Restart-Computer -ComputerName $ComputerName -ErrorVariable errmsg -ErrorAction SilentlyContinue
-    if ($errmsg -ne $null) {
-        # Error log: Restart failed
-        Write-Host "$errmsg" -ForegroundColor red
+
+    if ($errmsg -ne $null)
+    {
+        if ($errmsg -like "*other users logged on*")
+        {
+            Write-Host "$($Pc.'Computer Name') failed: $(@($errmsg -split ':')[1].Trim())" -ForegroundColor yellow
+        }
+        else {
+            Write-Host "$($Pc.'Computer Name') failed: $(@($errmsg -split ':')[1].Trim())" -ForegroundColor red
+        }
     }
     else {
-        # Success
         Write-Host "$ComputerName restarted." -ForegroundColor green
     }
 }
